+++ Testing/Source/MediaStorageAndFileFormat/Cxx/TestImageRegionReader1.cxx
@@ -82,7 +82,7 @@ static int TestImageRegionRead(const char* filename, bool verbose = false)
   const char *ref = gdcm::Testing::GetMD5FromFile(filename);
 
   char digest[33];
-  gdcm::Testing::ComputeMD5(buffer, len, digest);
+  gdcm::Testing::ComputeMD5(buffer, (unsigned long)len, digest);
   if( verbose )
     {
     std::cout << "ref=" << ref << std::endl;
